@@ -10,28 +10,28 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^17.3.8",
+    "@angular/animations": "^18.2.3",
     "@angular/cdk": "^16.2.8",
-    "@angular/common": "^17.3.8",
-    "@angular/compiler": "^17.3.8",
-    "@angular/core": "^17.3.8",
-    "@angular/forms": "^17.3.8",
+    "@angular/common": "^18.2.3",
+    "@angular/compiler": "^18.2.3",
+    "@angular/core": "^18.2.3",
+    "@angular/forms": "^18.2.3",
     "@angular/material": "^16.2.8",
-    "@angular/platform-browser": "^17.3.8",
-    "@angular/platform-browser-dynamic": "^17.3.8",
-    "@angular/router": "^17.3.8",
+    "@angular/platform-browser": "^18.2.3",
+    "@angular/platform-browser-dynamic": "^18.2.3",
+    "@angular/router": "^18.2.3",
     "@emailjs/browser": "^4.3.3",
     "ng-recaptcha": "^13.2.1",
     "ng2-pdf-viewer": "^10.3.0",
     "pdfjs-dist": "^4.2.67",
     "rxjs": "~7.8.0",
     "tslib": "^2.3.0",
-    "zone.js": "~0.14.5"
+    "zone.js": "~0.14.10"
   },
   "devDependencies": {
     "@angular-devkit/build-angular": "^18.2.3",
-    "@angular/cli": "~17.3.7",
-    "@angular/compiler-cli": "^17.3.8",
+    "@angular/cli": "~18.2.3",
+    "@angular/compiler-cli": "^18.2.3",
     "@types/jasmine": "~4.3.0",
     "jasmine-core": "~4.6.0",
     "karma": "~6.4.0",
